Moved UseCases to seperate package & made them implement UseCase interface #121
