+++ fakes-core/src/main/java/be/xplore/recordreplay/usecase/StubHandler.java
@@ -1,4 +1,4 @@
-package be.xplore.fakes.service;
+package be.xplore.recordreplay.usecase;
 
 import be.xplore.fakes.model.Response;
 import be.xplore.fakes.model.Stub;
