+++ recordreplay-usecases/build.gradle
@@ -1,5 +1,4 @@
 dependencies {
     implementation project(':okhttp-client')
     implementation project(':fakes-core')
-    testCompile group: 'junit', name: 'junit', version: '4.12'
 }
