+++ recordreplay-usecases/src/main/java/be/xplore/recordreplay/usecase/ForwardRequestUseCase.java
@@ -1,0 +1,26 @@
+package be.xplore.recordreplay.usecase;
+
+import be.xplore.fakes.model.Request;
+import be.xplore.fakes.model.Response;
+import be.xplore.fakes.model.Stub;
+import be.xplore.fakes.service.HttpClient;
+
+import java.util.Optional;
+
+public class ForwardRequestUseCase implements UseCase{
+
+    private final HttpClient httpClient;
+
+    public ForwardRequestUseCase(HttpClient httpClient) {
+        this.httpClient = httpClient;
+    }
+
+    Response forward(Request request) {
+        return httpClient.execute(request);
+    }
+
+    public Optional<Response> execute(Stub stub) {
+        return Optional.of(forward(stub.getRequest()));
+    }
+
+}
