+++ recordreplay-usecases/src/main/java/be/xplore/recordreplay/usecase/RecordReplayUseCase.java
@@ -1,4 +1,4 @@
-package be.xplore.usecases;
+package be.xplore.recordreplay.usecase;
 
 import be.xplore.fakes.model.Response;
 import be.xplore.fakes.model.Stub;
@@ -6,16 +6,14 @@ import be.xplore.fakes.util.Assert;
 
 import java.util.Optional;
 
-public class RecordReplayUseCase {
+public class RecordReplayUseCase implements UseCase {
 
-    private final ReplayUseCase replay;
-    private final ForwardRequestUseCase forward;
     private final RecordUseCase record;
+    private final ReplayUseCase replay;
 
-    public RecordReplayUseCase(ReplayUseCase replay, ForwardRequestUseCase forward, RecordUseCase record) {
-        this.replay = Assert.notNull(replay);
-        this.forward = Assert.notNull(forward);
+    public RecordReplayUseCase(ReplayUseCase replay, RecordUseCase record) {
         this.record = Assert.notNull(record);
+        this.replay = Assert.notNull(replay);
     }
 
     public Optional<Response> execute(Stub stub) {
@@ -23,7 +21,7 @@ public class RecordReplayUseCase {
         if (foundResp.isPresent()) {
             return foundResp;
         } else {
-            return record.execute(new Stub(stub.getRequest(), forward.execute(stub)));
+            return record.execute(stub);
         }
     }
 }
