+++ recordreplay-usecases/src/main/java/be/xplore/recordreplay/usecase/ReplayUseCase.java
@@ -1,4 +1,4 @@
-package be.xplore.usecases;
+package be.xplore.recordreplay.usecase;
 
 import be.xplore.fakes.model.Response;
 import be.xplore.fakes.model.Stub;
@@ -8,7 +8,7 @@ import be.xplore.fakes.service.RequestMatcher;
 import java.util.List;
 import java.util.Optional;
 
-public class ReplayUseCase {
+public class ReplayUseCase implements UseCase{
 
     private final Repository repository;
     private final List<RequestMatcher> matchers;
